Project complete
I used the tutorial of Krish Naik to learn how to use NLP to prepare the data for training a ML model. In this I experimented with TfIdf vectorisation and compared the accuracy rate for both BOW and TfIdf. I also used lemmatisation to increase the accuracy of the model trained on TfIdf.